hw3/streams/requirements.txt
faust-streaming==0.11.3
